회원 탈퇴시 작성한 글 먼저 삭제

+++ src/main/java/com/example/prj1be20231109/controller/BoardController.java
@@ -2,9 +2,13 @@ package com.example.prj1be20231109.controller;
 
 
 import com.example.prj1be20231109.domain.Board;
+import com.example.prj1be20231109.domain.Member;
 import com.example.prj1be20231109.service.BoardService;
+import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
+import lombok.Value;
 import org.apache.ibatis.annotations.Delete;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 ;import java.util.List;
@@ -18,14 +22,20 @@ public class BoardController {
 
     @PostMapping("add")
     public ResponseEntity add(
-            @RequestBody Board board
+            @RequestBody Board board,
+            @SessionAttribute(value = "login",required = false) Member login
     ){
+
+        if(login==null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
         if(!service.validate(board)){
             return ResponseEntity.badRequest().build();
         }
 
 
-       if(service.save(board)){
+       if(service.save(board,login)){
           return ResponseEntity.ok().build();
        }else {
           return ResponseEntity.internalServerError().build();
@@ -46,8 +56,19 @@ public class BoardController {
 
     @DeleteMapping("remove/{id}")
     public ResponseEntity remove(
-            @PathVariable Integer id
+            @PathVariable Integer id,
+            @SessionAttribute(value="login", required=false) Member login
     ){
+
+        if(login==null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build(); // 401
+        }
+
+
+        if(!service.hasAccess(id, login)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build(); // 403 권한
+        }
+
         if(service.remove(id)){
             return  ResponseEntity.ok().build();
         }else{
@@ -57,8 +78,18 @@ public class BoardController {
 
     @PutMapping("update")
     public ResponseEntity update(
-            @RequestBody Board board
+            @RequestBody Board board,
+            @SessionAttribute(value = "login",required = false) Member login
     ){
+        if (login==null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build(); //401
+        }
+
+
+        if(!service.hasAccess(board.getId(),login)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build(); // 403
+        }
+
         if(!service.validate(board)){
             return ResponseEntity.badRequest().build();
         }
