+++ src/main/java/com/example/prj1be20231109/controller/MemberController.java
@@ -2,6 +2,7 @@ package com.example.prj1be20231109.controller;
 
 import com.example.prj1be20231109.domain.Member;
 import com.example.prj1be20231109.service.MemberService;
+import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -72,9 +73,16 @@ public class MemberController {
     }
 
     @GetMapping
-    public ResponseEntity<Member> view(String id) {
-        // TODO : 로그인 했는 지? -> 안했으면 401
-        // TODO : 자기 정보인지? -> 아니면 403
+    public ResponseEntity<Member> view(String id
+    ,@SessionAttribute(value = "login",required = false)Member login
+    ) {
+        if(login==null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
+        if (!service.hasAccess(id,login)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
 
         Member member = service.getMember(id);
 
@@ -82,9 +90,18 @@ public class MemberController {
     }
 
     @DeleteMapping
-    public ResponseEntity delete(String id) {
-        // TODO : 로그인 했는 지? -> 안했으면 401
-        // TODO : 자기 정보인지? -> 아니면 403
+    public ResponseEntity delete(String id,
+     @SessionAttribute(value = "login", required = false) Member login
+     ) {
+        if (login==null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build(); //401
+        }
+
+        if(!service.hasAccess(id,login)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build(); //403
+        }
+
+
 
         if (service.deleteMember(id)) {
             return ResponseEntity.ok().build();
@@ -95,8 +112,17 @@ public class MemberController {
     }
 
     @PutMapping("edit")
-    public ResponseEntity edit(@RequestBody Member member) {
-        // TODO: 로그인 했는지? 자기정보인지?
+    public ResponseEntity edit(@RequestBody Member member,
+      @SessionAttribute(value = "login",required = false)Member login
+     ) {
+        if (login==null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build(); //401
+        }
+
+        if (!service.hasAccess(member.getId(),login)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build(); //403
+        }
+
 
         if (service.update(member)) {
             return ResponseEntity.ok().build();
@@ -115,6 +141,13 @@ public class MemberController {
             return ResponseEntity.ok().build();
         }else {
             return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+    }
+
+    @PostMapping("logout")
+    public void logout(HttpSession session){
+        if(session != null){
+            session.invalidate();
         }
     }
 }
